+++ meta/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to the Orbital Trading project will be documented in this file by the user, only. Gemini is never to modify or provide updates to this file, and must always treat this file as read-only. The user will always manually update this file.
 
+# [35.26] - 2026-2-13
+-  added time-cost to hull repair
+-  services screen polish
+-  added proportial hull stress multiplier to speed of travel
+-  updated fuel cost of event delays
+-  hull deterioration rate increased by 1%
+-  hull deterioration detached from time and attached to distance instead
+
 # [35.20] - 2026-2-12
 -  mission screen polish
 -  updated meta files